fixed deadlock/timeout issue in process job creation workflow

+++ inferno/lib/daemon.py
@@ -31,6 +31,7 @@ def unpickle_connection(pickled_connection):
     (func, args) = pickle.loads(pickled_connection)
     return func(*args)
 
+
 def run_rule_async(rule_name, automatic_cycle, settings, reply_to):
     setproctitle("inferno - %s" % rule_name)
     signal.signal(signal.SIGHUP, signal.SIG_IGN)
@@ -140,13 +141,17 @@ class InfernoDaemon(object):
             raise e
 
     def die(self, x=None, y=None):
-        print 'dying...'
-        try:
-            self.disco_ball.stopped = True
-            self.disco_ball.server.terminate()
-        except:
-            pass
-        os._exit(0)
+        pid = os.getpid()
+        if not self.disco_ball.stopped:
+            print 'dying... %d' % pid
+            try:
+                self.disco_ball.stopped = True
+                self.disco_ball.server.terminate()
+            except:
+                pass
+            os._exit(0)
+        else:
+            print 'dead... %d' % pid
 
     def start(self):
         signal.signal(signal.SIGTERM, self.die)
